requirements.txt
Django>=3.2
djangorestframework
django-notifications-hq
pytest
pytest-django
